refactor: Toast Ui 에러 핸들링 추가
